+++ charms/planner-operator/src/charm.py
@@ -7,7 +7,6 @@
 import dataclasses
 import json
 import logging
-import pathlib
 import typing
 
 import ops
@@ -61,14 +60,6 @@ class GithubRunnerPlannerCharm(paas_charm.go.Charm):
             self.on[PLANNER_RELATION_NAME].relation_broken,
             self._on_planner_relation_broken,
         )
-
-    def get_cos_dir(self) -> str:
-        """Get the COS directory for this charm.
-
-        Returns:
-            The COS directory.
-        """
-        return str((pathlib.Path(__file__).parent / "cos").absolute())
 
     def _create_app(self):
         """Patch _create_app to add OpenTelemetry environment variables."""