+++ docs/changelog.md
@@ -2,8 +2,12 @@
 
 # Changelog
 
-All notable changes to this project will be documented in this file.
+This changelog documents user-relevant changes to the Planner charm and Webhook gateway charm.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
+
+## 2026-04-13
+
+- add 5xx error logging to planner routes.